+++ R/Image_application.R
@@ -2,19 +2,8 @@
 #' @name apply_to_image
 #' @description
 #'   This function could apply the defined water cluster to corrected image files.
-#'   Should run \code{generate_param()} to generate a \code{res} list as an input
+#'   Should run \link{generate_param} to generate a \code{res} list as an input
 #'   of function \code{apply_to_image}
-#'
-#' @usage apply_to_image(input, res,
-#'   output_image=TRUE, output_resultpng=FALSE, output_imRrs.n=FALSE,
-#'   Chla_est=FALSE,
-#'   title.name = NULL, png_scale=50,
-#'   fn_memb="output_membership",
-#'   fn_cluster="output_cluster",
-#'   fn_imRrs.n='output_imRrs_normalized',
-#'   fn_truecolorpng='output_truecolor',
-#'   fn_Chla = 'output_Chla',
-#'   output_format='GTiff')
 #'
 #' @param input A \strong{raster} or a \strong{character} linking
 #'   to the raster file on the disk.
@@ -27,6 +16,9 @@
 #'   }
 #'   For the convenience, function \link{generate_param} supports to quickly
 #'     generate this \code{list}. See more in examples.
+#' @param color_palette The palette of cluster color. Default as \code{RdYlBu(res$K)}.
+#'   In \code{FCMm}, it could be \link{RdYlBu}, \link{Spectral}, \link{HUE} or other color values
+#'   with same length of cluster number.
 #' @param output_image Logical, whether to produce image files
 #' @param output_resultpng Logical, whether to produce png files
 #' @param output_imRrs.n Logical, whether to produce normalized Rrs files
@@ -87,13 +79,15 @@
 #' @family Fuzzy cluster functions
 #' 
 #' @import ggplot2
-#' @import raster
+#' @importFrom raster raster brick as.data.frame rasterFromXYZ stretch rasterFromXYZ 
+#'   crs crs<-
 #' @importFrom reshape2 melt
 #' @importFrom magrittr %>% %<>%
 #' @importFrom ggthemes theme_map
+#' @importClassesFrom raster Raster RasterBrick RasterStack
 #' 
-
-apply_to_image <- function(input, res,
+apply_to_image <- function(input, res, 
+                           color_palette = RdYlBu(res$K),
                            output_image=TRUE, output_resultpng=FALSE, output_imRrs.n=FALSE,
                            Chla_est=FALSE,
                            title.name = NULL, png_scale=50,
@@ -108,7 +102,10 @@ apply_to_image <- function(input, res,
     stop("The input of image file is missing!")
   if(missing(res))
     stop("FCM results is missing")
-
+  
+  if(length(color_palette) != res$K)
+    stop("The length of color_palette shoud be same with cluster number!")
+  
   message("Since we have not check the input image file,")
   message("  please make sure the wavelength of image file and cluster dataframe match correspondingly.")
   message("The normalization of spectra is default in thie version!")
@@ -124,7 +121,7 @@ apply_to_image <- function(input, res,
 
   if(im@data@nlayers!=length(res$FD$wv))
     stop("The band number of image file is different from wavelength length!")
-  imdf <- as.data.frame(im,na.rm=TRUE,xy=TRUE)
+  imdf <- raster::as.data.frame(im,na.rm=TRUE,xy=TRUE)
   
   x_name <- which(names(imdf) == "x")
   y_name <- which(names(imdf) == "y")
@@ -134,7 +131,7 @@ apply_to_image <- function(input, res,
   names(imRrs.raw) <- paste0("Rrs",wv)
   Rrs <- as.matrix(imRrs.raw)
   Area <- trapz(wv,Rrs)
-  Area <- as.data.frame(Area)
+  Area <- base::as.data.frame(Area)
   imRrs.n <- imRrs.raw
   for(i in 1:ncol(imRrs.raw))
     imRrs.n[,i] = imRrs.raw[,i] / Area
@@ -156,13 +153,13 @@ apply_to_image <- function(input, res,
   res.FCM <- apply_FCM_m(Rrs=imRrs.n, wavelength=wv, Rrs_clusters=v,
                          default.cluster=FALSE)
   res.im <- list()
-  res.im$u <- res.FCM$u %>% as.data.frame
+  res.im$u <- base::as.data.frame(res.FCM$u)
   res.im$cluster <- res.FCM$cluster
 
   # Save true color image
   rgb <- brick(input[[10]],input[[7]],input[[5]],input[[2]])
   rgb_stretch <- stretch(x=rgb, minv=0, maxv=255)
-  rgb_df <- as.data.frame(rgb_stretch,xy=TRUE)
+  rgb_df <- raster::as.data.frame(rgb_stretch, xy=TRUE)
   rgb_df <- data.frame(x=rgb_df$x, y=rgb_df$y,
                        n=rgb_df[,3],r=rgb_df[,4], g=rgb_df[,5],b=rgb_df[,6]) %>% na.omit
   p.truecolor=ggplot(data=rgb_df) +
@@ -208,7 +205,7 @@ apply_to_image <- function(input, res,
     message(paste0("Cluster map was generated, named: ", fn_cluster))
   }
   message("Plotting clusters ......")
-  cp <- RdYlBu(nrow(v))
+  cp <- color_palette
   cp.sub <- cp[(unique(res.im$cluster)) %>% sort]
   p.cluster <- ggplot() +
     geom_raster(data=im.cluster,aes(x=x,y=y,fill=as.character(res.im.cluster)),