+++ R/tools.R
@@ -1126,4 +1126,8 @@ Spectral <- grDevices::colorRampPalette(RColorBrewer::brewer.pal(11, "Spectral")
 #' @rdname Color_plattes
 RdYlBu <- grDevices::colorRampPalette(RColorBrewer::brewer.pal(11, "RdYlBu"))
 
-
+#' @export
+#' @rdname Color_plattes
+#' @importFrom scales hue_pal
+#' @importFrom farver encode_colour
+HUE <- scales::hue_pal()